% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npranov.R
\name{interaction_plot}
\alias{interaction_plot}
\title{Generate an Interactive Interaction Plot}
\usage{
interaction_plot(data, factor1, factor2, response)
}
\arguments{
\item{data}{Data frame containing the factors and response variable.}

\item{factor1}{First factor variable.}

\item{factor2}{Second factor variable.}

\item{response}{Numeric response variable.}
}
\value{
An interactive interaction plot using ggplot2 and plotly.
}
\description{
Generate an Interactive Interaction Plot
}
\examples{
interaction_plot(data, "Location", "Tribe", "Midichlorians")
}
