% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npranov.R
\name{krss}
\alias{krss}
\title{Kernel Ridge Regression and ANOVA-based Analysis}
\usage{
krss(response, A, B, n_perm = 1000)
}
\arguments{
\item{response}{Numeric vector (dependent variable).}

\item{A}{Factor variable (independent factor A).}

\item{B}{Factor variable (independent factor B).}

\item{n_perm}{Number of permutations for significance testing (default = 1000).}
}
\value{
A dataframe with p-values and significance levels for factor A, factor B, and their interaction.
}
\description{
Kernel Ridge Regression and ANOVA-based Analysis
}
\examples{
data <- data.frame(
  Location = factor(rep(c("Olympia", "Ventura"), 6)),
  Tribe = factor(rep(c("Jedi", "Sith"), 6)),
  Midichlorians = c(10, 4, 12, 5, 15, 4, 15, 9, 15, 11, 18, 12)
)
Krss(data$Midichlorians, data$Location, data$Tribe)
}
