% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npranov.R
\docType{package}
\name{npranov}
\alias{npranov-package}
\alias{npranov}
\title{Krss: Kernel Ridge Regression and ANOVA-based Analysis}
\description{
This package provides functions to perform kernel ridge regression, permutation-based significance testing,
and visualization of main and interaction effects.
}
\author{
\strong{Maintainer}: Group 03 \email{amiduyohan@gmail.com}

}
